1. Implemented some missing functionality int the breakpoints widget. Edit and delete breakpoint. 2. Undid some reduced font size in the register widget. 3. Made some changes in the interface to the debugger to enable messages between debugger windows. 4. Fixed some small bugs in register widget. 4. Added possibility to navigate up and down in the disassembler window. Note: I feel it is ready for a merge to trunk...

+++ main/gui_qt4/debugger/registerwidget.h
@@ -32,7 +32,7 @@ class TableListModel;
 // used to check for changes
 const extern long long int reg_cop1_fgr_64[32];
 const extern long long int reg[32];
-const extern long int reg_cop0[32];
+const extern unsigned int reg_cop0[32];
 const extern long long int hi;
 const extern long long int lo;
 const extern unsigned int ai_register;
@@ -82,7 +82,7 @@ class RegisterWidget : public QWidget, private Ui_RegisterWidget
         QString getCop1String(long long int);
         
         long long int gui_fantom_gpr_64[32];
-        long int gui_fantom_cop0_32[32];
+        unsigned int gui_fantom_cop0_32[32];
         long long int gui_fantom_cop1_64[32];
         long long int gui_fantom_hi;
         long long int gui_fantom_lo;
